refactor code and moved view components to views directory
